% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_songs}
\alias{demo_songs}
\title{Demo songs to load in as the \code{state} argument to \code{tr808()}}
\format{
An object of class \code{list} of length 3.
}
\usage{
demo_songs
}
\description{
Demo songs to load in as the \code{state} argument to \code{tr808()}
}
\keyword{datasets}
