% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_names}
\alias{sample_names}
\title{Names of all instrument samples}
\format{
An object of class \code{character} of length 11.
}
\usage{
sample_names
}
\description{
Names of all instrument samples
}
\keyword{datasets}
