% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet.R
\name{text_to_tr808_state}
\alias{text_to_tr808_state}
\title{Convert a text representation into a \code{tr808_state} object with a single pattern}
\usage{
text_to_tr808_state(txt, bpm = 96)
}
\arguments{
\item{txt}{Single string must contain 11 strings of length 16 characters consisting of
only \code{_} and \code{x}.  All other characters will be ignored.}

\item{bpm}{bpm. default: 96}
}
\value{
\code{tr808_state} object
}
\description{
The returned object can be loaded into the drum machine using
\code{tr808(state = ...)}
}
\examples{
text_to_tr808_state(
'#RStats #tr808r
 BD xx____xx__x_____
 SD ____x_______x___
 LT ________________
 MT xx______________
 HT ______xx__x__x__
 RS ________________
 CP xx_xx_xx_______x
 CB xx__x_xx_x_x_xx_
 CY ________________
 OH ________________
 CH x_xxx_xxx_xxx_xx'
)
}
