% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr808.R
\name{tr808}
\alias{tr808}
\title{TR-808}
\usage{
tr808(bpm = 96, width = 12, height = 8, state = NULL, ...)
}
\arguments{
\item{bpm}{beats per minute. default: 96}

\item{width, height, ...}{arguments passed to \code{eventloop::run_loop()}}

\item{state}{a \code{tr808_state} object or the path of a saved object in
an RDS file.  Note: Pressing 's' within the interactive drum machine
window will save the current state to a time stamped RDS file.}
}
\description{
TR-808
}
