% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr808.R
\name{tr808_func}
\alias{tr808_func}
\title{Main TR808 eventloop callback}
\usage{
tr808_func(event, mouse_x, mouse_y, event_env, ...)
}
\arguments{
\item{event, mouse_x, mouse_y, event_env, ...}{eventloop args}
}
\description{
Main TR808 eventloop callback
}
